Revert "Examen no lo encuentra lo saque"
This reverts commit 3b9ed9602e99c027de52490429806e7b6933c346.

+++ KendoDP2/Areas/Reclutamiento/Models/Respuesta.cs
@@ -12,7 +12,7 @@ namespace KendoDP2.Areas.Reclutamiento.Models
     {
 
         public string Comentario { get; set; }
-        public double Puntaje { get; set; }
+        public int Puntaje { get; set; }
 
         public int EvaluacionXFaseXPostulacionID { get; set; }
         public virtual EvaluacionXFaseXPostulacion EvaluacionXFaseXPostulacion { get; set; }
@@ -48,7 +48,7 @@ namespace KendoDP2.Areas.Reclutamiento.Models
     {
         public int ID { get; set; }
         public string Comentario { get; set; }
-        public double Puntaje { get; set; }
+        public int Puntaje { get; set; }
         public int EvaluacionXFaseXPostulacionID { get; set; }
         public int FuncionID { get; set; }
